Validate the method and destination
Method and destination validation are being added together because the
two go together. This will ensure that one of `--sd` and `--ssh` is
specified on the CLI (since the CLI doesn't currently enforce this
itself). Based on which method is picked, it will also validate that the
destination exists.

In the case of syncing via SSH, no actual validation of the destination
is done. I could check for a corresponding entry in the SSH config file
(this is how the current `retro` function does it), but this would
exclude the ability to sync to a known host name or IP address. I could
also try to ping the destination, but that is a lot of extra overhead
for successful calls and more or less the same overhead as letting the
sync fail for unsuccessful calls.