Standardize config loading and parsing
Both the `link` and the `retro` modules load and parse a configuration
file and store them in very similar structs. This moves them all of that
to a new module and replaces the existing structs with common ones.
Having most of the fields as optional simplifies this.

This change also makes it possible to encapsulate the logic for handling
multiple, one, or no destinations and extensions in methods on the
common structs. And they now have tests!