Change some of the exits to return Errs
This application is structured such that the main entry point returns an
exit code, 0 or 1, based on whether or not the CLI returns an `Ok` or an
`Err`. Ideally this would be the only place this happens. This removes
most of the others. The only one being left for now is the one that is
triggered when a required environment variable isn't set, but this both
simplifies the calling code and includes `or_exit` in its name to
indicate that it could do this.