Merge onion subcommand into link
The more I used this application, the more I was reminded that one of my
primary motivations for unifying things was so that I wouldn't have to
run two commands each time I backed up a new game (at least for the
systems that are playable on OnionOS). This merges the `onion`
subcommand into `link`. The latter will now call out to functions that
handle symlinking and copying for the various destinations. It will
gracefully handle trying to sync a system that can't be process through
either of the functions.

There's probably some cleanup that can be done but I wanted to get the
merge itself into a working state first.