Add subcommand to rename bin and cue files
When I back up CD-based games, they typically come out as bin and cue
files. Unless the dumper is able to identify the game, the files have
generic prefixes (`track` in the case of MPF, `cdrom` in the case of
RetroArch). This subcommand provides a way to mass rename the files by
replacing the common prefix with a new one. More importantly, it will
also do the same to the contents of the cue files, which list out the
related bin files.

Ultimately I'd like to update this to take in a path (e.g., a directory)
and figure out what the common prefix used for all files is, then handle
the rest from there.