Add subcommand to compress files
This command can be used to compress cue and iso files to chd. It takes
a directory as input and will find all cue and iso files in it. It
optionally takes a directory for where to store the output. It will skip
any files for which a chd already exists in the output directory.

Some day I'd like to be able to compress to other formats as well, such
as rvz.