Default to the directory name when renaming files
When renaming bin and cue files, they often live inside a folder that's
named the same as what I want to name the files. Rather than having to
do something like

    retro rename "Some Folder Name" "Some Folder Name"

This will allow me to omit the second argument and use the first for
both. This looks something like

    retro rename "Some Folder Name"

Much better. I've also made it smart enough to trim the trailing slash
that shows up when I use tab completion for the directory name.