Prefer slices to vectors
Using vectors requires ownership of the memory. Using slices borrows the
memory. While I still need to use vectors for the CLI and when
populating structs from TOML, almost every other place that I use a
vector can be simplified by using a slice instead.

One place that feels like a sideways step is when determining the list
of systems. Once the `if` block is done, the memory used for
`config.get_system_names` is freed, so the memory needs to be claimed
beforehand.