Stream output when compressing
`chdman` streams its output, updating the completion percentage as it
goes. When the output is captured and printed only after the command
completes, this real-time information is lost.

This adds a `stream_output` function that works similarly to
`capture_output`, except, it will allow the running process to write
directly to stdout. It will also exit with the process's exit code if
it's non-zero.