Add an m3u alias for the playlist command
I like to use meaningful names for my commands, but `playlist` is long
compared to `m3u`. And the latter is how I often mentally think of the
command. This adds an alias so that either will work. It's being added
as a visible alias so that it's discoverable in the help text.