Add a verbosity flag to the command line
The output produced by this tool can sometimes be a bit much. I've
decided to add a verbosity flag to help control this. Fortunately,
there's a [clap_verbosity_flag] crate that does most of the heavy
lifting for me. It requires switching from `println` and `eprintln` to a
logger, but this is probably for the best. I've stripped out the
logger's prefix and done by best to match the current experience,
adjusting a few log levels here and there.

I plan to follow this up with a change that will take advantage of the
flag to control the streaming output as well.