Use the verbosity flag to suppress chdman output
The output from chdman can be useful, but it isn't always something I
want to see. This uses the verbosity flag to either stream the output
(the current behavior), suppress the output, or display a completion
message (the default).

A change to the API of a single `dispatch` function is being introduced
here. I don't think I like the idea of them not all having the same
arguments anymore, but I don't yet need it for any of the other commands
so I'm going to leave it for now, and revisit this if/when I come up
with other uses.