Check if logging level is enabled
Rather than passing around the logging level and doing numerical
comparisons, the `log_enabled` macro can be used to check if the
messages logged at the specified level will be logged. This not only
makes the intent clearer, it also means I can restore the signature of
this one instance of `dispatch` to match all of the others again.