Check for existence of correct system source
When linking games, this tool checks for the existence of each system's
source before trying to sync it. While merging the code for linking
games for Onion into the code that handles my general game linking, I
renamed a variable. Unfortunately, I missed one reference to the
variable: the one used for the actual check. This updates that to avoid
a panic when I try to link a system that doesn't have a source.