Make some output a little easier to read
When a `PathBuf` is formatted using the `Debug` trait, it escapes
characters like apostrophes. This can cause it to look a little off at
times. This switches some logging messages to use the `Display` trait
instead. The one downside is that the whole value is no longer wrapped
in quotes, but that's easy enough to add myself if it ends up bothering
me.