Link games regardless of log level
It turns out the logging macros are smart enough to not execute any
calls being passed in as arguments if logs of the specified level would
not be logged. This means that calling games won't be linked if the log
level is anything less than `info`. This moves the `capture_output`
calls to separate lines with assignments.