Don't symlink games that are already symlinked
This will check for and resolve symlinks that already exist in the
destination folder when linking games and skip any that already
reference the game being linked.

The log message about creating a link is being promoted from `info` to
`warn` with logs about skipping being handled as `info`. For
consistency, the output from `rsync` for Onion games is also being
promoted to `warn`.