Allow renaming with or without a trailing slash
When I made the rename subcommand use the source's name as the default
in cfbdd27, I made it work with trailing slashes in the name.
Unfortunately I seem to have forgotten to test it without a trailing
slash. `strip_suffix` can't be unwrapped when the suffix isn't found. I
thought it would return the original value unmodified. This lead to a
panic when the source didn't include a trailing slash and no specific
destination value was provided. This recreates that logic using
`unwrap_or`.