Use more specific names for configuration
When I first started this, the only external configuration needed was
for the destinations written to by the link subcommand. While that isn't
entirely changing, I'd like to make the link subcommand more
configurable so that any number of destinations can be synced. The
existing configs will stay the same, but their names are sometimes too
generic. I'd like to reserve names like `Config` and `load_config` for a
higher-level config file. I'm updating their names to indicate that
they're related to the destinations used by the link subcommand.