Make link destinations configurable
When I first wrote this, I did it with the idea of syncing games to two
locations, one via symlinking and one via copying. This was because the
latter is synced onto a Windows machine so that it could be read by
[Skraper] which was configured to generate output for [OnionOS]. Windows
couldn't handle symlinks and so the files needed to by copied. [OnionOS]
can now handle all of the media and metadata itself so I intend to
remove [Skraper] from my workflow. With this change, I can switch to
symlinks for all of my files.

At this point, it seems unnecessary to have distinct ways to handle my
general game and [OnionOS] use cases. I've decided to reuse the former
for the latter. To achieve this, I'm adding a system-level configuration
that can be used to configure subcommands, starting with link. The
config will contain the source (formerly hardcoded to `$RETRO_BACKUPS`)
and any destinations (formerly hardcoded as `$RETRO_GAMES` and
`$ONION_GAMES`). The config will support both paths and environment
variables that expand to paths. The [OnionOS]-specific logic of
supporting multiple destinations for a system is being preserved.

One change of note here: [OnionOS] (because of the underlying Miyoo
firmware) expects folders to be named in all uppercase. Everywhere else
I use my games, they are lowercase. I added in some code to convert the
system name to uppercase for the destination if no explicit destination
was provided. Because I run this utility on macOS, it was fine with this
for my general game sync since those folders all already existed in
their lowercase form. I'm removing this behavior in favor of requiring
uppercase names to be specified explicitly.

[onionos]: https://onionui.github.io
[skraper]: https://www.skraper.net