Publish the crate from GitHub
Rather than building the binary and attaching it to a GitHub release,
this switches to an action that will publish it directly to crates.io.
While I won't be able to just download a pre-built binary, I can use
`cargo` to install the latest version which can probably be automated a
little better.