Only publish crate when source of manifest changes
This adds a step to the publish action that should prevent the action
from trying to publish the crate if no changes are made within `src` or
to `Cargo.toml`. I may want to extend this to `Cargo.lock`, but I'm
going to wait and see if Dependabot ever sends a pull request for just
that file. The motivation here was to allow me to make GitHub- and
linting-related changes without publish an unchanged crate.