Update the version number when publishing crate
The publish action is being updated to include steps to set the version
number so that each release will be treated as a newer one. Calver will
still be used, with the current year and month being grabbed in the
action and the number of commits being used as the patch number. This
won't reset each month, but the whole thing is arbitrary and only exists
so that `cargo install` will see and install new releases.

As part of this change, the version number in `Cargo.toml` is being set
to `0.0.0`. This is being done to make it easier to identify it with
`sed`, but it also has the advantage that using a version from source
will have a different version number than a version installed by
`cargo`.

This approach is based on [a workflow I found on GitHub][workflow].

[workflow]: https://github.com/Emoun/duplicate/blob/1b504d3229aedcea1e34712d69d0e4f4e95e3ac0/.github/workflows/rust.yml#L74-L87