Exit the publish workflow early
Rather than doing a bunch of work and skipping only the `cargo publish`
step in the publish workflow, this will exit out of the workflow early
if no relevant changes were found. I've been having trouble with the
workflow not publishing the package when expected and I'm hoping that
exiting early instead of just skipping a step will help.