Load config files through confy
[Command Line Applications in Rust][book] recommends using [confy] for
working with config files. Taking a look at it, it's a pretty
straightforward replacement for what I've been doing and leads to less
but more flexible code. There are some improvements that can be made
here, but I have to start somewhere.

With this change I can remove [toml] as a direct dependency.

[book]: https://rust-cli.github.io/book/in-depth/config-files.html
[confy]: https://docs.rs/confy/
[toml]: https://docs.rs/toml/