Use default config but support override
Rather than requiring the `$RETRO_CONFIG` environment variable, this
will check for it and fall back to the default confy config file if it
isn't found. In order to support this, a new function, `get_from_env`,
is being added. Unlike `env_or_exit`, it will handle a missing
environment variable gracefully. `env_from_exit` is being renamed to
`get_from_env_or_exit` to keep the two functions consistent.

With this change, there is increased potential for the app to run with
an empty config and not do much, at least in terms of linking games. Not
only do I need to document the structure of the config, but I should
also consider exposing a way to modify it through the app itself.