Support compressing to DVD
Some emulators support CHDs in DVD format. This adds a flag to the
compress command that will cause it to use `createdvd` instead of
`createcd`. As this is emulator-specific, I'd like this to happen
without having to pass in an argument each time the command is invoked,
but for now this will do.