Support compress-specific config
This introduces a special config for the compress command. In a
departure from how the config for the link command works, this structure
of this config is defined as part of the `compress` module. A new
function, `load_config_recursively`, is being added to find a
`retro.toml` file in a directory or one of its ancestors.

Unlike `load_config` and `load_link_destination_config`,
`load_config_recursively` returns a generic type; the caller must
specify the type. I'd like to adapt this for the link command's config
as well, but I might run into limitations with the presence of a config
used by another command. If that doesn't go well, I may need to rethink
the approach here.