Move logic for finding a file recursively to utils
The code that finds a file recursively may have uses outside of just
loading a config so I decided to move it out of
`load_config_recursively` and create a new function,
`find_file_recursively`. This has the added benefit of making the logic
easier to test.

While adding the tests I noticed that the `utils` module's tests were in
the middle of the file. They are being moved to the end. When running
the tests, I noticed that the naming was different between the `config`
and `utils` modules, so the latter's tests are being renamed to match.