Allow manually skipping the publish workflow
This adds a check to the publish workflow for the presence of
`[skip publish]` in the commit message. Its presence will cause the
workflow to exit early and not publish the crate. This was motivated by
some changes that only modify tests. There's no reason to publish a new
version of the crate, but I don't want to use flag to skip all workflows
so that the tests themselves will still run.