Treat skipped workflows as successful
When I added logic to the publish workflow to skip it, I decided to have
it return a non-zero exit code. While I like seeing that the job don't
run completely, seeing that it failed is a bit misleading. Plus it can
cause me to miss an actual failure with the job.