Rename load_config to load_global_config
There is a bit of a mismatch between the purpose of `load_config` and
`load_config_recursively`. The latter will look for a config file in the
specified location (or its ancestors); the former loads the
application-wide, a.k.a. global, config. The former is being renamed to
show this difference. It opens the door to adding a `load_config` later
on that works like `load_config_recursively` but without checking
ancestors.