Add a dispatch trait to commands
This introduces an unbound trait that adds a `dispatch` method to
commands. While this increases the boilerplate needed when defining a
command, it reduces some of it when dispatching to them.

I'd love to simplify things inside `cli.rs` even more, but since each
command's arguments are of a different type, I'll probably need a macro
that generates the match branches for me accomplish that.