Add clean subcommand to remove broken links
Sometimes I move files around or rename them and end up with broken
symlinks. This adds a clean subcommand that will look for such links and
remove them. The subcommand accepts a flag to tell it whether or not to
perform a dry run instead of actually removing the files.