Strip leading zero from month in version number
When getting the currently month, single digit values are prefixed with
a zero. I _thought_ `$((month+0))` was stripping the zero. It turns out,
however, that Cargo treats the value as a number, not a string, so it
was ignoring the zero.

`08` is treated as an octal value rather than as a decimal one. So once
August rolled around, the publish workflow began to fail because of

```
invalid leading zero in minor version number
```

This updates the approach to stripping the leading zero. Since there is
no way to test the workflow (and manually testing the old approach
worked), hopefully this works.

Reference: https://stackoverflow.com/a/62757839