Break up find_files
`find_files` is designed to find files in a directory that match one or
more extensions. This functionality is useful. It will continue to live
on as `find_files_with_extension`. The inner logic to find all files --
what drives the extension filter -- is being split into a separate
function. It will receive the freed up `find_files` name.

This function isn't immediately useful but it will become valuable as I
make some changes to the config logic.