Add initial code to parse dat files
I'd like to be able to organize my games. I'd like to use the names
established by databases such as No-Intro and Redump. This is the first
step in getting there. This adds a module that can parse the dat files
provided by these databases. The structures used for the data is
currently based on a file from No-Intro. I suspect it will need to be
modified when tested with a file from Redump (e.g., multiple files for a
game).